+++ src/linux_python_utils/logging/__init__.py
@@ -1,6 +1,7 @@
 """Module de logging."""
 
 from linux_python_utils.logging.base import Logger
+from linux_python_utils.logging.console_logger import ConsoleLogger
 from linux_python_utils.logging.file_logger import FileLogger
 from linux_python_utils.logging.security_logger import (
     SecurityEvent,
@@ -10,6 +11,7 @@ from linux_python_utils.logging.security_logger import (
 
 __all__ = [
     "Logger",
+    "ConsoleLogger",
     "FileLogger",
     "SecurityEvent",
     "SecurityEventType",